tried +closest intersection but its kind of broken fixed ppatch

+++ proj2/rayd2.h
@@ -35,8 +35,8 @@ class rayd2
  private:
   float rSphere(SlVector3 &ray,float* sOrigin); //intRsect sphere
 
-  int rTri(SlVector3 &ray,float* p); //inteRsect Triangle
-  int rTri(SlVector3 &ray,SlVector3* p);
+  float rTri(SlVector3 &ray,float* p); //inteRsect Triangle
+  float rTri(SlVector3 &ray,SlVector3* p);
   
   /*--given values--*/
   string m_ofile;
